Final Part 1.

+++ DatePickerHint/Controllers/StudentController.cs
@@ -72,9 +72,13 @@ namespace DatePickerHint.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var student = await _context.Students.FindAsync(id);
+            if (student == null) return NotFound();
+
             _context.Students.Remove(student);
             await _context.SaveChangesAsync();
+
             return RedirectToAction(nameof(Index));
         }
+
     }
 }
